detector/requirements.txt
fastapi
uvicorn
torch
torchvision
timm
Pillow
